Timestamp Index refactor
